% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelinevis.R
\name{timelinevis}
\alias{timelinevis}
\title{Create a timelinevis widget}
\usage{
timelinevis(items, width = NULL, height = NULL)
}
\arguments{
\item{items}{a dataframe that will be converted into a timelinejs DataSet.
Each row will be one timeline entry.
See vis.js documentation for valid column names}
}
\description{
Send a dataframe containing timeline entries to the javascript widget
}
\examples{
items=data.frame(content=c("item1","item2"), start=c("2013-04-02","2013-04-14"))
timelinevis(items)

}

